core: autopilot-manager: Isolate mavlink_proxy
* Isolate mavlink_proxy type definitions and exceptions in its own
  module
