+++ core/services/ardupilot_manager/exceptions.py
@@ -63,37 +63,5 @@ class NoDefaultFirmwareAvailable(RuntimeError):
     """Default firmware file is not available."""
 
 
-class EndpointCreationFail(RuntimeError):
-    """Failed to add endpoint."""
-
-
-class EndpointDeleteFail(RuntimeError):
-    """Failed to delete endpoint."""
-
-
-class EndpointUpdateFail(RuntimeError):
-    """Failed to update endpoint."""
-
-
-class MavlinkRouterStartFail(RuntimeError):
-    """Failed to initiate Mavlink router."""
-
-
-class NoMasterMavlinkEndpoint(ValueError):
-    """No master Mavlink endpoint set."""
-
-
-class EndpointAlreadyExists(ValueError):
-    """Mavlink endpoint already exists."""
-
-
-class DuplicateEndpointName(ValueError):
-    """Another mavlink endpoint with same name already exists."""
-
-
-class EndpointDontExist(ValueError):
-    """Given Mavlink endpoint do not exist."""
-
-
 class NoPreferredBoardSet(RuntimeError):
     """No preferred board is set yet."""
