+++ core/services/ardupilot_manager/mavlink_proxy/AbstractRouter.py
@@ -8,14 +8,14 @@ from typing import Any, List, Optional, Set, Type
 
 from loguru import logger
 
-from exceptions import (
+from mavlink_proxy.Endpoint import Endpoint
+from mavlink_proxy.exceptions import (
     DuplicateEndpointName,
     EndpointAlreadyExists,
     EndpointDontExist,
     MavlinkRouterStartFail,
     NoMasterMavlinkEndpoint,
 )
-from mavlink_proxy.Endpoint import Endpoint
 
 
 class AbstractRouter(metaclass=abc.ABCMeta):
